[feat/#94] 추천 API 구현 (#95)
* feat: 추천 게시글 조회 API 구현

* feat: 추천 재생성 API 구현

* chore: 사용하지 않는 리포지토리 메서드 제거

* refactor: ES 쿼리 모듈화

* fix: Nested 필드 참조 시 잘못된 필드명 참조하는 로직 해결

+++ src/main/java/com/techfork/domain/recommendation/config/RecommendationProperties.java
@@ -34,9 +34,9 @@ public class RecommendationProperties {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class EmbeddingWeights {
-        private Double title = 0.4;
-        private Double summary = 0.4;
-        private Double content = 0.2;
+        private Float title = 0.4f;
+        private Float summary = 0.4f;
+        private Float content = 0.2f;
     }
 
     @Getter
